# docs/swagger.yaml
definitions:
  api.CreateSubRequest:
    properties:
      end_date:
        example: 12-2025
        type: string
      price:
        example: 100
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  api.CreateSubResponse:
    properties:
      message:
        example: Подписка успешно создана
        type: string
    type: object
  api.DeleteSubRequest:
    properties:
      start_date:
        example: 07-2025
        type: string
    type: object
  api.DeleteSubResponse:
    properties:
      message:
        example: Подписка успешно удалена
        type: string
    type: object
  api.ErrorResponse:
    properties:
      error:
        example: Некорректный формат идентификатора пользователя
        type: string
    type: object
  api.SubResponse:
    properties:
      end_date:
        example: 12-2025
        type: string
      price:
        example: 100
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
    type: object
  api.TotalCostRequest:
    properties:
      total_from:
        example: 07-2025
        type: string
      total_to:
        example: 09-2025
        type: string
    type: object
  api.TotalCostResponse:
    properties:
      message:
        example: 'Общая стоимость подписки Yandex Plus за указанный период составила:
          300'
        type: string
    type: object
  api.UpdateSubRequest:
    properties:
      new_end_date:
        example: 12-2025
        type: string
      new_price:
        example: 100
        type: integer
    type: object
  api.UpdateSubResponse:
    properties:
      message:
        example: Уровень подписки повышен и уже действует. Дата окончания подписки
          изменена
        type: string
    type: object
info:
  contact: {}
paths:
  /subscriptions:
    post:
      consumes:
      - application/json
      description: Создает новую подписку для пользователя
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/api.CreateSubRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Подписка успешно создана
          schema:
            $ref: '#/definitions/api.CreateSubResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Подписка уже существует
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Создать подписку
      tags:
      - subscriptions
  /users/{user_id}/subscriptions:
    get:
      description: Возвращает список подписок для указанного user_id
      parameters:
      - description: UUID пользователя
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список подписок
          schema:
            items:
              $ref: '#/definitions/api.SubResponse'
            type: array
        "400":
          description: Некорректный UUID пользователя или service_name
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Получить подписки пользователя
      tags:
      - subscriptions
  /users/{user_id}/subscriptions/{service_name}:
    delete:
      consumes:
      - application/json
      description: Удаляет подписку пользователя по дате начала действия
      parameters:
      - description: UUID пользователя
        in: path
        name: user_id
        required: true
        type: string
      - description: Название сервиса
        in: path
        name: service_name
        required: true
        type: string
      - description: Дата начала подписки для удаления
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.DeleteSubRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Подписка успешно удалена
          schema:
            $ref: '#/definitions/api.DeleteSubResponse'
        "400":
          description: Некорректные данные запроса
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Подписка не найдена
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Возвращает список подписок для указанного user_id
      parameters:
      - description: UUID пользователя
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список подписок
          schema:
            items:
              $ref: '#/definitions/api.SubResponse'
            type: array
        "400":
          description: Некорректный UUID пользователя или service_name
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Получить подписки пользователя
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет уровень и/или дату окончания подписки пользователя
      parameters:
      - description: UUID пользователя
        in: path
        name: user_id
        required: true
        type: string
      - description: Название сервиса
        in: path
        name: service_name
        required: true
        type: string
      - description: Новые данные подписки
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateSubRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об обновлении подписки
          schema:
            $ref: '#/definitions/api.UpdateSubResponse'
        "400":
          description: Некорректные данные запроса
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Подписка не найдена
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Обновить подписку
      tags:
      - subscriptions
  /users/{user_id}/subscriptions/{service_name}/total:
    post:
      consumes:
      - application/json
      description: Рассчитывает общую стоимость подписки за период
      parameters:
      - description: UUID пользователя
        in: path
        name: user_id
        required: true
        type: string
      - description: Название сервиса
        in: path
        name: service_name
        required: true
        type: string
      - description: Период total_from / total_to
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.TotalCostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение с суммой подписки
          schema:
            $ref: '#/definitions/api.TotalCostResponse'
        "400":
          description: Некорректные данные запроса
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Подсчитать общую стоимость подписки
      tags:
      - subscriptions
swagger: "2.0"